Implement `MessageFieldObjectGenerator` and `MessageOneofObjectGenerator`.
